refactor: move alpha_beta function
